// backend/src/types/controlIva.ts
export interface Supplier {
  corr?: number;
  cod_proveedor: string;
  cod_emp?: number;
  nom_proveedor: string;
  dir_proveedor?: string;
  cod_dept?: number;
  cod_muni?: number;
  telefono?: string;
  registro?: string;
  nit_proveedor?: string;
  giro?: string;
  exento?: number;
  exterior?: number;
  activo?: number;
  tama?: string;
  pais?: string;
  cuenta_contable?: string;
  nombre_cuenta?: string;
  con_credito?: number;
  excede_credito?: number;
  limite_credito?: number;
  con_retencion?: number;
  con_percepcion?: number;
  identificacion_excluidos?: string;
  deducible?: number;
  nom_dept?: string;
  nom_muni?: string;
}
// ...
export interface PaginatedResult<T> {
  data: T[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}